Authentication

+++ graphql/resolvers/bookings.js
@@ -3,8 +3,9 @@ import Event from '../../models/event.js';
 import { transformEvent, transformBooking } from './merge.js';
 
 export default {
-  bookings: async () => {
+  bookings: async (args, req) => {
     try {
+      if (!req.isAuth) throw new Error('UnAuthorized!');
       const bookings = await Booking.find({});
       return bookings.map((booking) => {
         return transformBooking(booking);
@@ -14,8 +15,10 @@ export default {
     }
   },
 
-  bookEvent: async (args) => {
+  bookEvent: async (args, req) => {
     try {
+      if (!req.isAuth) throw new Error('UnAuthorized!');
+
       const fetchedEvent = await Event.findOne({ _id: args.eventId });
       const booking = new Booking({
         user: '5fe566fd1e793b197c4b5a73',
@@ -28,6 +31,7 @@ export default {
     }
   },
   cancelBooking: async (args) => {
+    if (!req.isAuth) throw new Error('UnAuthorized!');
     const booking = await Booking.findById(args.bookingId).populate('event');
     const event = transformEvent(booking.event);
     await Booking.deleteOne({ _id: args.bookingId });
